feat(auth): clients can authenticate using URL parameters
This change provides basic functionality for authentication
based on URL parameters, using a auth token.

Clients must provide both username and password URL parameter,
although username is ignored as of now. The password will be
treated as an auth token, i.e. the user is derived from the
token as well.