feat(donation): allow people to donate money
We expose a new API endpoint at POST /donation/. When sent information on

- IBAN
- BIC
- name
- email (optional)
- amount
- and an optional message

then an email is sent to both deSEC (donation@desec.io) and the donor (if
an email address was provided). The deSEC email contains a SEPA direct
debit file for import into Hibiscus/Jameica banking software. This file
is using Nils' Creditor Identifier, so please be considerate.

The donations do not relate to user accounts. The IBAN is shortened before
stored into the database for security reasons (the email of course contains
the full IBAN).

The API server DOES NOT keep track of IP addresses to prevent abuse, as
using the IP address for prosecution of fraud is pointless anyways.

!!! This commit contains database migrations !!!