fix(mysql): configure mysql to support 4-byte utf8 chars
The utf8 charset of mysql does not support 4byte utf8 chars like 汉语 or
emojis. This commit changes the charset in use to utf8mb4, which does
support 4-byte chars. As the key length in mysql/mariadb is limited to
a certain number of *bytes* (not chars), this decreases the maximum key
length that is legal to 191 chars. It was hence necessary to provide a
migration that can update old databases as well as changing all old
migrations to only use 191 chars per key field.

Bonus: this commit fixes a typo in the mysql password env variable.