feat(network): BREAKING CHANGE internal network refactoring
- make IPv4 internal net prefix configurable
- restructure internal networks into several smaller ones, to restrict
  container visibility

NOTE: This commit changes some of the internal container IPs, even if you
set the new IPv4 prefix variable to its previous implicit  value (172.16).
You need to update the MariaDB grant tables as well  as the IP address in
the dbmaster supermaster table.
Also, if you change the IPv4 internal net prefix variable in the future,
the tables will need to be updated manually as well. See README.