feat(pdns): command lord to NOTIFY master immediately on update
This decreases DNS update time at the nsmaster to approx. one 1 second.

Adopted unit tests, as most tests now need to access the last-but-one
request to make their assertions.

Added a unit test to see if our code PUTs on the /notify endpoint of
the pdns API.

Calling notify_zone immediately after creating a zone results in a
500 error at the pdns API. For discussion, see also #40.