feat(api): RRset model, and endpoints to read and write RRsets
Endpoints (all accept HEAD, OPTIONS):

 1. `domains/<name>/rrsets/`: return all RRsets, or create an RRset. Accepts GET, POST.
 2. `domains/<name>/rrsets/<subname>?.../`: return RRsets for given subname (all types). Accepts GET.
 3. `domains/<name>/rrsets/<type>/`: return RRsets for given type (all subnames). Accepts GET.
 4. `domains/<name>/rrsets/<subname>?.../<type>/`: return RRset for given subname and type. Accepts GET, PATCH, PUT.

In case 4, <name>, <subname>, and <type> cannot be changed. The
only two remaining data members to change are the ttl and the
record values. A pdns update is only performed if at least one
of those is changed,

This commit also adds _ to the regex used for domains in urls.py.