20170418 rrsets endpoint (#50)
feat(api): REST APi to read and write RRsets

Endpoints:

 1. `domains/<name>/rrsets/`: return all RRsets, or create an RRset.
    Accepts GET (optionally with `type` and `subname` query parameters),
    POST.
 2. `domains/<name>/rrsets/<subname>?.../<type>/`: return RRset for
    given subname and type. Accepts GET, PATCH, PUT.

<name>, <subname>, and <type> cannot be changed.  A pdns update is only
performed if at least one of those is changed.  RRsets of type SOA,
RRSIG, DNSKEY, NSEC3PARAM are managed automatically and cannot be
tinkered with.

This commit also introduces the management command sync-from-pdns which
is used to import RRsets from the pdns database into the database
of our Django API app ("local database").   A list of domain names can
be provided as arguments.  If no names are provided, all domains that
are locally known are imported.  This way, it is possible to place
additional domain data in the pdns database, and if a corresponding
domain is not found locally, we do not try to import any associated
RRsets from pdns.  Any previously present local RRsets pertaining to the
imported domains are deleted.  In other words, this import is really a
one-way sync operation.

We provide tests for all new functionality.

For more details, please check the docs supplied within this commit.