feat(www): BREAKING CHANGE autogenerate self-signed certs if needed
This commit introduces autogeneration of self-signed certificates for
www. A list of DNS names for which certificates will be generated
can be found in www/autocert.sh.

This commit is a BREAKING CHANGE as it changes the way certificate
files are expected to be named.

To not use autogenerated certificates, place the following files
in your certificate folder (path from ENV):

- checkipv4.dedyn.${DESECSTACK_DOMAIN}
- checkipv6.dedyn.${DESECSTACK_DOMAIN}
- checkip.dedyn.${DESECSTACK_DOMAIN}
- dedyn.${DESECSTACK_DOMAIN}
- desec.${DESECSTACK_DOMAIN}
- update6.dedyn.${DESECSTACK_DOMAIN}
- update.dedyn.${DESECSTACK_DOMAIN}
- www.dedyn.${DESECSTACK_DOMAIN}
- www.desec.${DESECSTACK_DOMAIN}

(all ending in .cer and .key).