fix(api): make sure unlock() won't overwrite unmanaged pdns domains
When unlocking a user, we now try POST'ing domains to pdns if the user
created them in our API after being locked, and we do not catch the
exception that occurs when that domain is not new to pdns (like, when
the domain was created manually in the pdns database).

Previously, the exception was ignored (because the same function was
used to sync contents of existing domains to pdns, and these two cases
could not be distinguished).

We already have a test (testCantPostDomainAlreadyTakenInPdns) to make
sure that we error out correctly when the above pdns error occurs
without being caught.  I extended the test
testSuspendedUpdatesDomainCreation to also check that unlocking
triggers POST on the pdns domain endpoint, and that the exception is
actually thrown.