fix(api): improve validation of donation input data
Previously, IBAN and BIC were normalized and stripped from
whitespace only for creating the direct debit attachment file.
If the user entered spaces like '   DE123455....', those were
stored in the database. Along with the 6-character cutoff, this
was not desirable.

Space stripping is now done in the Donation serializer so that
it applies to both what's stored and to what's put in the direct
debit file.

Also, e2e tests were adapted to verify the presence of the masked
IBAN.