fix(api): disallow tinkering with OPT RRset
OPT is a reserved record type with special meaning and is not supposed
to exist in a zone. Currently, if an OPT RRset is created, the zone may
break.

Until PowerDNS ensures that RRsets of type OPT cannot created, let's
make sure on this level.

PowerDNS issue: https://github.com/PowerDNS/pdns/issues/6441