feat(ns): improve provisioning on nsmaster, fixes #190
Previously, zones were provisioned on nsmaster using the supermaster
feature of pdns, followed by a NOTIFY from nslord, triggering an AXFR
by nsmaster.

This mechanism relied on various configuration in database and pdns
containers, and further was subject to a race condition: If a user
created a new zone and immediately deleted the supermaster hostname
from the NS RRset, the AXFR would fail and the domain would not be
provisioned on nsmaster.

This commit removes the dependency on specific supermaster hostnames,
reduces configuration complexity and makes sure that new domains are
created on and AXFR'ed by nsmaster regardless of circumstances.