fix(nslord): improve SOA record settings
This commit changes the SOA MNAME field from ns1.desec.io to value
that does not refer to a specific nameserver. The field value is
irrelevant to the public, and this way the overspecific name is
removed.

(RFC1996 says that a master server "is named in the zone's SOA MNAME
field and optionally by an NS RR," that is, the MANME hostname does
not really need to be an NS. RFC7719 confirms this for more modern
times.)

It also increases the default TTL for SOA records to 3600: There's
no need for external resolvers to forget our SOA records every minute.