fix(api): POST/PATCH/PUT'ing RRset payload of invalid structure
Before this change, we read from request.data before validation, assuming that
it is a dict. Things went wrong if it was a list of dicts (sent to an
RRsetDetail endpoint accidentally), or a string, or something else.

Instead of adding extra checks, this commit removes usage of request.data
before validation. As a consequence, PUT on RRsetDetail now requires
specifying subname and type. (DRF requires to specify all fields for PUT.)