feat(www): remove general rate limit
tl;dr: We can cope with high load on static assets, and there seems
to be no scenario where an IP-based rate-limit would help. We should
be monitoring latency and bandwith.

I conducted tests in production from the same datacenter and concluded
that a rate limit is not needed for static content.

Tests (n: number of requests, c: concurrency):
- non-stress latency: ~20ms
- n=50000, c=100, 1.6KB file: 100% in 46s, 6 MB/s, latency ~90ms
- n=7500, c=100, 540KB file: 100% in 57s, 60-80 MB/s, latency ~750ms
- n=50000, c=1000, 1.6KB file: 99.9% in 46s, 6 MB/s, latency ~90ms

In the last case, there were 40 failures during SSL handshake. Rate
limiting happens after establishing the TLS connection, so it would
not have helped.