feat(www): recalibrate API IP rate limit
We now have 30/s for /api/ and 4/m for the dyn update endpoints,
per IP. There do not seem to be plausible use cases that exceed
this.

On the other hand, we have one user who causes about 6.5% of all
update requests on some days, sometimes sending up to 15 in one
second. Applying this limit will get rid of most of these
requests, and maybe alert the user of their misconfiguration.

We currently do not have any users who update more than four dyn
domains at the same time. If someone wants to to that, they can
add `sleep 1` or something like that to their cronjob.