feat(www): add montitoring (via stub_status and exporter sidecar)
This adds monitoring to nginx by converting the output of the
stub_status module to metrics in Prometheus format by means of
NGINX Inc.'s official Prometheus exporter:
https://github.com/nginxinc/nginx-prometheus-exporter

More advanced monitoring is possible using VTS exporter. As this
requires another module from a third-party author, I deemed it
better to only do this if we find it to be necessary. Reference:
https://hnlq715.github.io/nginx-vts-exporter/