fix(webapp): highlight table rows by setting state
Manipulation of the DOM is the exclusive job of vue.js. Modification
of the DOM directly yields to unexpected outcome, like in this case
where row highlighting could be messed up by changing table sort
order (the highlight would remain in place even if the row moves
away).

I think it's acceptable for now to keep the 'successFade' DOM
manipulation, as it is very short term.