fix(webapp): increase width of subname column in RR set list
The problem with the automatic scaling of column width in v-data-table
is that all our contents (even the ones you cannot edit) are input
element. Input elements must have a defined width -- as opposed to
other elements, they do not grow with content. Now which width
should we assign? I couldn't find a satisfactory solution, so this
just adjusts the pre-defined column widths a bit.

closes #476