fix(test): squash migrations 1 through 31
This commit /manually/ squashes migrations pertaining the
user.email field. This was necessary due to upstream removal of
CIEmailField, resulting in an extra index in the api database
that made migration 0031 fail. The manual fix was to move all
changes pertaining user.email to migration 0001. Among all
affected migrations, the one with the highest number is 0031.

This commit also Django-squashes migrations 1 through 31
(including) to make clear that database versions in this range
aren't supported anymore.

If you maintain a desec-stack deployment and this breaking
change is a problem, please contact support@desec.io. Thanks!