feat(www): double amount of worker processes to 128
This change is expected to decrease the amount of
502 responses sent by nginx due to "no live upstreams".

Memory usage is expected to increase by factor 2 from 9GiB
to 18GiB, or less due to shared memory resources.

Perhaps a better solution would be to increase the
number of threads, however for that the PDNS change tracker
needs to be upgraded to support multi-threading.